Update sorting and unit test of sortPrice
