+++ source/__tests__/unit_tests.js
@@ -5,7 +5,7 @@
  */
 
 /* global test, expect */
-const indexFunctions = require('../scripts/index.js').TestEnvironment
+const sortFunctions = require('../scripts/sortFunctions.js')
 
 // Dummy test to suppress warning
 test('adds 1 + 2 to equal 3', () => {
@@ -15,11 +15,11 @@ test('adds 1 + 2 to equal 3', () => {
     expect(operand1 + operand2).toBe(expected);
 });
 
-// TODO: add directive to the sorting function script
+// Test sortPrice function
 test('sort coffee cards by ascending price', () => {
     const coffeeCardObject1 = {
         "str_drink_name": "chestnut latte",
-        "int_drink_price": "12.5",
+        "float_drink_price": "12.5",
         "time_purchase_date": "2022-12-14",
         "str_purchase_location": "Starbucks",
         "img_drink_image": 3,
@@ -34,7 +34,7 @@ test('sort coffee cards by ascending price', () => {
     } 
     const coffeeCardObject2 = {
         "str_drink_name": "Mocha",
-        "int_drink_price": "5.21",
+        "float_drink_price": "5.21",
         "time_purchase_date": "2022-12-14",
         "str_purchase_location": "Starbucks",
         "img_drink_image": 3,
@@ -49,7 +49,7 @@ test('sort coffee cards by ascending price', () => {
     } 
     const coffeeCardObject3 = {
         "str_drink_name": "Frappucino",
-        "int_drink_price": "10",
+        "float_drink_price": "10.0",
         "time_purchase_date": "2022-12-14",
         "str_purchase_location": "Starbucks",
         "img_drink_image": 3,
@@ -68,13 +68,12 @@ test('sort coffee cards by ascending price', () => {
     coffeeCardObject2["time_creation_time"] = d.toLocaleTimeString();
     coffeeCardObject3["time_creation_time"] = d.toLocaleTimeString();
 
-    coffeeCards = []
+    coffeeCards = [coffeeCardObject1, coffeeCardObject2, coffeeCardObject3];
 
-    coffeeCards.push(coffeeCardObject1);
-    coffeeCards.push(coffeeCardObject2);
-    coffeeCards.push(coffeeCardObject3);
     
-    coffeeCards.sort(indexFunctions.sortPrice);
-    expect(coffeeCards[1]["int_drink_price"]).toBe("10")
+    coffeeCards.sort(sortFunctions.sortPrice);
+    expect(coffeeCards[0]["float_drink_price"]).toBe("5.21");
+    expect(coffeeCards[1]["float_drink_price"]).toBe("10.0");
+    expect(coffeeCards[2]["float_drink_price"]).toBe("12.5");
 
 })
