+++ source/index.html
@@ -79,7 +79,7 @@
         <button id = "cancel" type = "reset">CANCEL</button>
         
         <input type = "text" id="str_drink_name" placeholder="name of drink" name = "str_drink_name" required/>
-        <input type = "text"id="int_drink_price" placeholder = "price" name = "int_drink_price" required/>
+        <input type = "text"id="float_drink_price" placeholder = "price" name = "float_drink_price" required/>
         <!-- <input type = "text"id="time_purchase_date" placeholder = "date purchased" name = "time_purchase_date" required/> -->
         <input type="date" id="time_purchase_date"  name = "time_purchase_date" required/>
         <input type = "text" id="str_purchase_location" placeholder = "location" name = "str_purchase_location" required/>
